requirements.txt
Flask
python-docx
pandas